+++ CHANGELOG.md
@@ -5,6 +5,7 @@ the [Keep a CHANGELOG](http://keepachangelog.com/) principles.
 
 ## Unreleased
 
+- Fixed: Initial field values were not showing in Dialog component
 - Added: Field::hideInPreview() to only show form field in form and not in form preview.
 - Added: Table ArchiveRowAction preset.
 - Deprecated: hideIfEmpty(). Use hideInPreviewIfEmpty() instead to only hide field in preview when empty.